# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de EPIs.",
        "title": "SaaS EPI API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Suporte API",
            "url": "http://www.radaptech.com.br",
            "email": "suporte@seusaas.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/cadastro-funcionario": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cadastra um novo funcionário no sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Funcionários"
                ],
                "summary": "Cadastrar novo funcionário",
                "parameters": [
                    {
                        "description": "Dados do funcionário",
                        "name": "funcionario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.FuncionarioInserir"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Matrícula/Dados duplicados",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/funcionario/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove um funcionário pelo ID",
                "tags": [
                    "Funcionários"
                ],
                "summary": "Deletar funcionário",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do funcionário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sem Conteúdo"
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Atualiza os dados de um funcionário existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Funcionários"
                ],
                "summary": "Atualizar funcionário",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do funcionário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Novos dados do funcionário",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateFuncionarioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/funcionario/{matricula}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna os detalhes de um único funcionário",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Funcionários"
                ],
                "summary": "Buscar por matrícula",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matrícula do funcionário",
                        "name": "matricula",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Funcionario_Dto"
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/funcionarios": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna uma lista paginada de funcionários",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Funcionários"
                ],
                "summary": "Listar todos",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Página",
                        "name": "pagina",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Quantidade por página",
                        "name": "quantidade",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FuncionarioPaginado"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/funcionarios/completo": {
            "get": {
                "security": [
                    {
                        "BearerAut": []
                    }
                ],
                "description": "Retorna uma lista detalhada de todos os funcionários",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Funcionários"
                ],
                "summary": "Listar funcionários completos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.FuncionarioCompletoDto"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
        "/funcionarios/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna resumo dos funcionários para o dashboard",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Funcionários"
                ],
                "summary": "Dashboard de funcionários",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.FuncionarioDashbord"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/helper.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "helper.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.DepartamentoDto": {
            "type": "object",
            "properties": {
                "departamento": {
                    "type": "string",
                    "example": "Recursos Humanos"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "model.EntregaDoFuncionarioDto": {
            "type": "object",
            "properties": {
                "assinatura_digital": {
                    "type": "string"
                },
                "data_entrega": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "itens": {
                    "description": "\u003c-- AQUI ELA!",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.ItemEntregueDto"
                    }
                }
            }
        },
...
        "model.EpiResponse": {
            "type": "object",
            "properties": {
                "alerta_minimo": {
                    "type": "integer"
                },
                "ca": {
                    "type": "string"
                },
                "descricao": {
                    "type": "string"
                },
                "fabricante": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                },
                "protecao": {
                    "$ref": "#/definitions/model.TipoProtecaoDto"
                },
                "validade_ca": {
                    "type": "string"
                }
            }
        },
...
        "model.FuncaoDto": {
            "type": "object",
            "properties": {
                "cargo": {
                    "type": "string"
                },
                "departamento": {
                    "$ref": "#/definitions/model.DepartamentoDto"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.FuncionarioCompletoDto": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "entregas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EntregaDoFuncionarioDto"
                    }
                },
                "funcao": {
                    "$ref": "#/definitions/model.FuncaoDto"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "model.FuncionarioDashbord": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "model.FuncionarioInserir": {
            "type": "object",
            "required": [
                "cpf",
                "id_departamento",
                "id_funcao",
                "nome"
            ],
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "id_departamento": {
                    "type": "integer",
                    "minimum": 1
                },
                "id_funcao": {
                    "type": "integer",
                    "minimum": 1
                },
                "nome": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 3
                }
            }
        },
        "model.Funcionario_Dto": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "funcao": {
                    "$ref": "#/definitions/model.FuncaoDto"
                },
                "id": {
                    "type": "integer"
                },
                "matricula": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "model.ItemEntregueDto": {
            "type": "object",
            "properties": {
                "epi": {
                    "$ref": "#/definitions/model.EpiResponse"
                },
                "id": {
                    "type": "integer"
                },
                "quantidade": {
                    "type": "integer"
                },
                "tamanho": {
                    "$ref": "#/definitions/model.TamanhoDto"
                }
            }
        },
...
        "model.TamanhoDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "tamanho": {
                    "type": "string"
                }
            }
        },
...
        "model.TipoProtecaoDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateFuncionarioRequest": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "id_departamento": {
                    "type": "integer"
                },
                "id_funcao": {
                    "type": "integer"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
...
        "service.FuncionarioPaginado": {
            "type": "object",
            "properties": {
                "funcionario": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Funcionario_Dto"
                    }
                },
                "pagina": {
                    "type": "integer"
                },
                "paginaFinal": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}